Move config into submodule
